test the performance using the gtest framework
