+++ deepes/test/torch_sin_config.prototxt
@@ -6,7 +6,7 @@ gaussian_sampling {
 
 optimizer {
   type: "Adam",
-  base_lr: 0.005,
+  base_lr: 0.001,
   momentum: 0.9,
   beta1: 0.9,
   beta2: 0.999,
